// src/app/recipe-manager/models/recipe.ts
export class Recipe {
    recipeViewName: string;
    recipeDataFileName: string;
    recipePdfFileName: string;
    ingredients: string;
    directions: string;
    prepInstructions: string;
    nutritionInfo: string;
    calories: number; // Add more data later, create nutInfo model
    rating: number; // 1-10 ??

    constructor() {}
}